Cleanup, part 1.
